requirements.txt
streamlit>=1.28.0
openai>=1.3.0
pillow>=10.0.0
reportlab>=4.0.0
requests>=2.31.0
numpy>=1.24.0
fastapi>=0.110.0
uvicorn[standard]>=0.23.0
gunicorn>=21.2.0
PyPDF2>=3.0.0

